Nodes that link to websites
I have coloured node that will link to websites I have profiles on. I
have spcified attributes for the force of the diagram. This changes link
distance, friction and charge. Charge is a very low value - it might
need to be made slightly less low. I commented out by avatar since it is
not being loaded properly... Still need to fix centre node to be bigger.
Need to put images inside the nodes